[docs] Updated FeaturePopover example
