+++ website/docs/components/feature-popover/feature-popover-code.md
@@ -4,12 +4,14 @@ fileSource: feature-popover
 tabs: Design('feature-popover'), A11y('feature-popover-a11y'), API('feature-popover-api'), Example('feature-popover-code'), Changelog('feature-popover-changelog')
 ---
 
-## Animation example
+## Basic example
+
+To see the `FeaturePopover` again, reload the page.
 
 ::: sandbox
 
 <script lang="tsx">
-  export Demo from './examples/animation_example.tsx';
+  export Demo from './examples/basic_example.tsx';
 </script>
 
 :::
